Confirmation on Delete deck action

+++ src/pages/Decks/MyDecks/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {View, Text, Button, FlatList, TouchableOpacity} from 'react-native';
+import {View, Text, Button, FlatList, TouchableOpacity, Alert} from 'react-native';
 import DeckDao from '../../../dao/DeckDao';
 
 export default class MyDecks extends React.Component {
@@ -21,8 +21,22 @@ export default class MyDecks extends React.Component {
     }
 
     deleteDeck = async (deck) => {
-        await new DeckDao().deleteDeck(deck.id);
-        this.loadDecks(); 
+        Alert.alert( 
+            'Delete Deck', 
+            'Are you sure you want to delete this deck ? All the cards in this deck will be deleted.',
+            [
+              { text: 'Cancel', onPress: () => console.log('Cancel Pressed')},
+              { 
+                  text: 'OK', 
+                  onPress: async () => {                
+                    await new DeckDao().deleteDeck(deck.id);
+                    this.loadDecks(); 
+                    } 
+              }
+            ],
+            { cancelable: true }
+          );
+
     }
     
     render(){
